apps/api/pyproject.toml
[project]
name = "kyc-api"
version = "0.1.0"
description = "Conversational KYC multi-agent API"
requires-python = ">=3.11,<3.13"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "pydantic>=2.9",
  "pydantic-settings>=2.6",
  "python-dotenv>=1.0",
  "httpx>=0.27",
  "sqlalchemy[asyncio]>=2.0",
  "asyncpg>=0.29",
  "alembic>=1.13",
  "langgraph>=0.2.50",
  "langgraph-checkpoint-postgres>=2.0",
  "psycopg[binary,pool]>=3.2",
  "langfuse>=2.50,<3",
  "qdrant-client>=1.12",
  "pypdf>=5.0",
  "pymupdf>=1.24",
  "python-multipart>=0.0.12",
  "deepface>=0.0.93",
  "tf-keras>=2.17",
  "Pillow>=10.4",
  "numpy>=1.26,<2",
]

[dependency-groups]
dev = [
  "pytest>=8.3",
  "pytest-asyncio>=0.24",
  "httpx>=0.27",
  "ruff>=0.7",
  "mypy>=1.13",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
